+++ hugot_ort_test.go
@@ -162,6 +162,38 @@ func TestTokenClassificationPipelineValidationORT(t *testing.T) {
 	tokenClassificationPipelineValidation(t, session)
 }
 
+// BERT-base NER tests
+
+func TestTokenClassificationBertNERORT(t *testing.T) {
+	opts := []options.WithOption{options.WithOnnxLibraryPath(onnxRuntimeSharedLibrary)}
+	session, err := NewORTSession(opts...)
+	checkT(t, err)
+	defer func(session *Session) {
+		destroyErr := session.Destroy()
+		checkT(t, destroyErr)
+	}(session)
+	tokenClassificationPipelineBertNER(t, session)
+}
+
+func TestTokenClassificationBertNERORTCuda(t *testing.T) {
+	if os.Getenv("CI") != "" {
+		t.SkipNow()
+	}
+	opts := []options.WithOption{
+		options.WithOnnxLibraryPath("/usr/lib64/onnxruntime-gpu/libonnxruntime.so"),
+		options.WithCuda(map[string]string{
+			"device_id": "0",
+		}),
+	}
+	session, err := NewORTSession(opts...)
+	checkT(t, err)
+	defer func(session *Session) {
+		destroyErr := session.Destroy()
+		checkT(t, destroyErr)
+	}(session)
+	tokenClassificationPipelineBertNER(t, session)
+}
+
 // Zero shot
 
 func TestZeroShotClassificationPipelineORT(t *testing.T) {